ractor::actor

Module actor_cell

ActorCell is reference counted actor which can be passed around as needed

This module contains all the functionality around the ActorCell, including the internal properties, ports, states, etc. ActorCell is the basic primitive for references to a given actor and its communication channels

Structs

ActorCell

An ActorCell is a reference to an Actor’s communication channels and provides external access to send messages, stop, kill, and generally interactor with the underlying Actor process.

Enums

ActorStatus

ActorStatus represents the status of an actor’s lifecycle

Constants

ACTIVE_STATES

Actor states where operations can continue to interact with an agent