ractor::actor::actor_ref

Struct ActorRef

pub struct ActorRef<TMessage> { /* private fields */ }

An ActorRef is a strongly-typed wrapper over an ActorCell to provide some syntactic wrapping on the requirement to pass the actor’s message type everywhere.

An ActorRef is the primary means of communication typically used when interfacing with super::Actors

Implementations

impl<TMessage> ActorRef<TMessage>

pub fn get_cell(&self) -> ActorCell

Retrieve a cloned ActorCell representing this ActorRef

pub fn notify_supervisor_and_monitors(&self, evt: SupervisionEvent)

Notify the supervisor and all monitors that a supervision event occurred. Monitors receive a reduced copy of the supervision event which won’t contain the crate::actor::BoxedState and collapses the crate::ActorProcessingErr exception to a String

evt - The event to send to this crate::Actor’s supervisors

impl<TMessage> ActorRef<TMessage>
where TMessage: Message,

pub fn send_message( &self, message: TMessage, ) -> Result<(), MessagingErr<TMessage>>

Send a strongly-typed message, constructing the boxed message on the fly

message - The message to send

Returns [Ok(())] on successful message send, [Err(MessagingErr)] otherwise

pub fn where_is(name: ActorName) -> Option<ActorRef<TMessage>>

Try and retrieve a strongly-typed actor from the registry.

Alias of crate::registry::where_is

impl<TMessage: Message> ActorRef<TMessage>

pub fn get_derived<TFrom>(&self) -> DerivedActorRef<TFrom>
where TMessage: From<TFrom>, TFrom: TryFrom<TMessage>,

Constructs the DerivedActorRef for a specific type from ActorRef. This allows an actor which handles either a subset of the full actor’s messages or a convertable type, allowing hiding of the original message type through implementation of From conversion.

Returns a DerivedActorRef where the message type is convertable to the ActorRef’s original message type via From. In order to facilitate MessagingErr::SendErr the original message type also needs to be reverse convertable via TryFrom to the TFrom type.

This method will panic if a send error occurs, and the returned message cannot be converted back to the TFrom type. This should never happen, unless conversions are created incorrectly.

impl<TKey, TMsg> ActorRef<FactoryMessage<TKey, RetriableMessage<TKey, TMsg>>>
where TKey: JobKey, TMsg: Message,

pub fn submit_retriable_job( &self, job: Job<TKey, TMsg>, strategy: MessageRetryStrategy, ) -> Result<(), Box<MessagingErr<FactoryMessage<TKey, RetriableMessage<TKey, TMsg>>>>>

When you’re talking to a factory, which accepts “retriable” jobs, this convenience function sets up the retriable state for you and converts your job to a retriable equivalent.

job: The traditional Job which will be auto-converted into a Job of RetriableMessage for you however the accepted field, if set, will be dropped as RetriableMessages do not support accepted replies, except on the first iteration

strategy: The MessageRetryStrategy to use for this retriable message

Returns the result from the underlying cast operation to the factory’s ActorRef.

impl<TMessage> ActorRef<TMessage>
where TMessage: Message,

pub fn cast(&self, msg: TMessage) -> Result<(), MessagingErr<TMessage>>

Alias of cast

pub async fn call<TReply, TMsgBuilder>( &self, msg_builder: TMsgBuilder, timeout_option: Option<Duration>, ) -> Result<CallResult<TReply>, MessagingErr<TMessage>>
where TMsgBuilder: FnOnce(RpcReplyPort<TReply>) -> TMessage, TReply: Send + 'static,

Alias of call

pub fn call_and_forward<TReply, TForwardMessage, TMsgBuilder, TFwdMessageBuilder>( &self, msg_builder: TMsgBuilder, response_forward: &ActorRef<TForwardMessage>, forward_mapping: TFwdMessageBuilder, timeout_option: Option<Duration>, ) -> Result<JoinHandle<CallResult<Result<(), MessagingErr<TForwardMessage>>>>, MessagingErr<TMessage>>
where TReply: Send + 'static, TMsgBuilder: FnOnce(RpcReplyPort<TReply>) -> TMessage, TForwardMessage: Message, TFwdMessageBuilder: FnOnce(TReply) -> TForwardMessage + Send + 'static,

Alias of call_and_forward

impl<TMessage> ActorRef<TMessage>
where TMessage: Message,

Add the timing functionality on top of the crate::ActorRef

pub fn send_interval<F>(&self, period: Duration, msg: F) -> JoinHandle<()>
where F: Fn() -> TMessage + Send + 'static,

Alias of send_interval

pub fn send_after<F>( &self, period: Duration, msg: F, ) -> JoinHandle<Result<(), MessagingErr<TMessage>>>
where F: FnOnce() -> TMessage + Send + 'static,

Alias of send_after

pub fn exit_after(&self, period: Duration) -> JoinHandle<()>

Alias of exit_after

pub fn kill_after(&self, period: Duration) -> JoinHandle<()>

Alias of kill_after