ractor::actor

Module derived_actor

DerivedActorRef wraps an ActorCell to send messages that can be converted to its accepted type using From. It represents a subset of the messages supported by the original actor.

Structs

DerivedActorRef

DerivedActorRef wraps an ActorCell to send messages that can be converted into its accepted type using From. DerivedActorRef allows to create isolation between actors by hiding the actual message type.