ractor::actor::derived_actor

Struct DerivedActorRef

pub struct DerivedActorRef<TFrom> { /* private fields */ }

DerivedActorRef wraps an ActorCell to send messages that can be converted into its accepted type using From. DerivedActorRef allows to create isolation between actors by hiding the actual message type.

Example

// In this example the actor is a ghost kitchen which can accept orders of different types,
// representing multiple virtual restaurants at once. More specifically, it can accept
// pizza or sushi orders.
//
// Derived actor allows to hide the order message accepted by the kitchen actor, therefore
// we can pass the ref to other components without creating a direct dependency on the
// kitchen actor. We can easily replace or split the kitchen actor without affecting
// other components communicating with it.
use ractor::{Actor, ActorProcessingErr, ActorRef, DerivedActorRef, Message};

// First we define order types
struct PizzaOrder {
    topping: String,
}

struct SushiOrder {
    r#type: String,
    quantity: usize,
}

// The order message which is actually sent to the kitchen actor can be either pizza or sushi
enum Order {
    Pizza(PizzaOrder),
    Sushi(SushiOrder),
}

// Implementing conversion methods from different order types to the actual order
impl From<PizzaOrder> for Order {
    fn from(value: PizzaOrder) -> Self {
        Order::Pizza(value)
    }
}

impl TryFrom<Order> for PizzaOrder {
    type Error = String;

    fn try_from(value: Order) -> Result<Self, Self::Error> {
        match value {
            Order::Pizza(order) => Ok(order),
            _ => Err("Order has invalid type".to_string()),
        }
    }
}

impl From<SushiOrder> for Order {
    fn from(value: SushiOrder) -> Self {
        Order::Sushi(value)
    }
}

impl TryFrom<Order> for SushiOrder {
    type Error = String;

    fn try_from(value: Order) -> Result<Self, Self::Error> {
        match value {
            Order::Sushi(order) => Ok(order),
            _ => Err("Order has invalid type".to_string()),
        }
    }
}

#[cfg(feature = "cluster")]
impl Message for Order {
    fn serializable() -> bool {
        false
    }
}

struct Kitchen;

#[cfg_attr(feature = "async-trait", ractor::async_trait)]
impl Actor for Kitchen {
    type Msg = Order;
    type State = ();
    type Arguments = ();

    async fn pre_start(
        &self,
        _myself: ActorRef<Self::Msg>,
        _: (),
    ) -> Result<Self::State, ActorProcessingErr> {
        Ok(())
    }

    async fn handle(
        &self,
        _myself: ActorRef<Self::Msg>,
        message: Self::Msg,
        _state: &mut Self::State,
    ) -> Result<(), ActorProcessingErr> {
        match message {
            Order::Pizza(order) => {
                println!("Preparing pizza with topping {}", order.topping);
            }
            Order::Sushi(order) => {
                println!(
                    "Preparing {} sushi of type {}",
                    order.quantity, order.r#type
                );
            }
        }
        Ok(())
    }
}

async fn example() {
    let (kitchen_actor_ref, kitchen_actor_handle) = Actor::spawn(None, Kitchen, ()).await.unwrap();
     
    // derived actor ref can be passed to the pizza restaurant actor which accepts pizza orders from delivery apps
    let pizza_restaurant: DerivedActorRef<PizzaOrder> = kitchen_actor_ref.get_derived();
    pizza_restaurant.send_message(PizzaOrder {
        topping: String::from("pepperoni"),
    }).expect("Failed to order pizza");

    // same way, we can also get a derived actor ref which can only accept sushi orders
    let sushi_restaurant: DerivedActorRef<SushiOrder> = kitchen_actor_ref.get_derived();
    sushi_restaurant.send_message(SushiOrder {
        r#type: String::from("sashimi"),
        quantity: 3,
    }).expect("Failed to order sushi");

    kitchen_actor_handle.await.unwrap();
}

Implementations

impl<TFrom> DerivedActorRef<TFrom>

pub fn send_message(&self, message: TFrom) -> Result<(), MessagingErr<TFrom>>

Casts the message to the target message type of ActorCell and sends it

message - The message to send

Returns [Ok(())] on successful message send, [Err(MessagingErr)] otherwise

pub fn get_cell(&self) -> ActorCell

Retrieve a cloned ActorCell representing this DerivedActorRef

impl<TMessage> DerivedActorRef<TMessage>
where TMessage: Message,

pub fn cast(&self, msg: TMessage) -> Result<(), MessagingErr<TMessage>>

Alias of cast

pub async fn call<TReply, TMsgBuilder>( &self, msg_builder: TMsgBuilder, timeout_option: Option<Duration>, ) -> Result<CallResult<TReply>, MessagingErr<TMessage>>
where TMsgBuilder: FnOnce(RpcReplyPort<TReply>) -> TMessage, TReply: Send + 'static,

Alias of call

impl<TMessage> DerivedActorRef<TMessage>
where TMessage: Message,

Add the timing functionality on top of the crate::ActorRef

pub fn send_interval<F>(&self, period: Duration, msg: F) -> JoinHandle<()>
where F: Fn() -> TMessage + Send + 'static,

Alias of send_interval

pub fn send_after<F>( &self, period: Duration, msg: F, ) -> JoinHandle<Result<(), MessagingErr<TMessage>>>
where F: FnOnce() -> TMessage + Send + 'static,

Alias of send_after

pub fn exit_after(&self, period: Duration) -> JoinHandle<()>

Alias of exit_after

pub fn kill_after(&self, period: Duration) -> JoinHandle<()>

Alias of kill_after