ractor::concurrency

Module tokio_primitives

Concurrency primitives based on tokio

Functions

interval

Build a new interval at the given duration starting at now

sleep

Sleep the task for a duration of time

spawn

Spawn a task on the executor runtime

spawn_local

Spawn a task on the executor runtime which will not be moved to other threads

spawn_named

Spawn a (possibly) named task on the executor runtime

timeout

Execute the future up to a timeout

Type Aliases

Duration

A duration of time

Instant

An instant measured on system time

Interval

An asynchronous interval calculation which waits until a checkpoint time to tick

JoinHandle

Represents a task JoinHandle

JoinSet

A set of futures to join on, in an unordered fashion (first-completed, first-served)

Attribute Macros

test

Marks async function to be executed by runtime, suitable to test environment. This macro helps set up a Runtime without requiring the user to use Runtime or Builder directly.