ractor::factory

Module discard

Discard handler managing when jobs are discarded

Enums

DiscardMode

The discard mode of a factory

DiscardReason

Reason for discarding a job

DiscardSettings

If a factory supports job discarding (loadshedding) it can have a few configurations which are defined in this enum. There is

Traits

DiscardHandler

Trait defining the discard handler for a factory.

DynamicDiscardController

Controls the dynamic concurrency level by receiving periodic snapshots of job statistics and emitting a new concurrency limit