ractor::factory

Module factoryimpl

Factory definition

Structs

Factory

Factory definition.

FactoryArguments

Arguments for configuring and starting a Factory actor instance.

FactoryArgumentsBuilder

Use builder syntax to set the required parameters and finish by calling the method Self::build().

FactoryState

State of a factory (backlogged jobs, handler, etc)