ractor::factory::job

Struct JobBuilder

pub struct JobBuilder<TKey, TMsg, ___State = __JobBuilderInitialState>
where
    TKey: JobKey,
    TMsg: Message,
{ /* private fields */ }

Use builder syntax to set the required parameters and finish by calling the method Self::build().

Implementations

impl<TKey, TMsg, __Key, __Msg, __Options, __Accepted> JobBuilder<TKey, TMsg, (__Key, __Msg, __Options, __Accepted)>
where TKey: JobKey, TMsg: Message,

pub fn build(self) -> Job<TKey, TMsg>
where __Key: IntoSet<TKey, JobBuilder__key>, __Msg: IntoSet<TMsg, JobBuilder__msg>, __Options: IntoSet<Option<JobOptions>, JobBuilder__options>, __Accepted: IntoSet<Option<RpcReplyPort<Option<Job<TKey, TMsg>>>>, JobBuilder__accepted>,

Finishes building and returns the requested object.

pub fn key(self, value: TKey) -> <Self as __JobBuilderSetMember>::Key
where __Key: IsUnset,

The key of the job

pub fn msg(self, value: TMsg) -> <Self as __JobBuilderSetMember>::Msg
where __Msg: IsUnset,

The message of the job

pub fn maybe_options( self, value: Option<JobOptions>, ) -> <Self as __JobBuilderSetMember>::Options
where __Options: IsUnset,

Same as Self::options, but accepts an Option as input. See that method’s documentation for more details.

pub fn options( self, value: JobOptions, ) -> <Self as __JobBuilderSetMember>::Options
where __Options: IsUnset,

The job’s options, mainly related to timing information of the job

Default = JobOptions::default()

pub fn maybe_accepted( self, value: Option<RpcReplyPort<Option<Job<TKey, TMsg>>>>, ) -> <Self as __JobBuilderSetMember>::Accepted
where __Accepted: IsUnset,

Same as Self::accepted, but accepts an Option as input. See that method’s documentation for more details.

pub fn accepted( self, value: RpcReplyPort<Option<Job<TKey, TMsg>>>, ) -> <Self as __JobBuilderSetMember>::Accepted
where __Accepted: IsUnset,

If provided, this channel can be used to block pushes into the factory until the factory can “accept” the message into its internal processing. This can be used to synchronize external threadpools to the Tokio processing pool and prevent overloading the unbounded channel which fronts all actors.

The reply channel return None if the job was accepted, or [Some(Job)] if it was rejected & loadshed, and then the job may be retried by the caller at a later time (if desired).

Default = None