ractor::factory

Module queues

Queue implementations for Factories

Structs

DefaultQueue

A simple, no-priority queue

PriorityQueue

A queue with NUM_PRIORITIES priorities

Enums

StandardPriority

Basic 5-category priority definition. This is probably flexible enough for most use-cases

Traits

Priority

Priority trait which denotes the usize value of a Priority

PriorityManager

The PriorityManager is responsible for extracting the job priority from a given job’s key (TKey). Additionally in some scenarios some jobs may be non-discardable, i.e. can be enqueued regardless of the backpressure status of the factory. This is also responsible for determining if a job can be loadshed.

Queue

Implementation of backing queue for factory messages when workers are all busy