ractor::factory::queues

Trait PriorityManager

pub trait PriorityManager<TKey, TPriority>:
    Send
    + Sync
    + 'static
where
    TKey: JobKey,
    TPriority: Priority,
{
    // Required methods
    fn is_discardable(&self, job: &TKey) -> bool;
    fn get_priority(&self, job: &TKey) -> Option<TPriority>;
}

The PriorityManager is responsible for extracting the job priority from a given job’s key (TKey). Additionally in some scenarios some jobs may be non-discardable, i.e. can be enqueued regardless of the backpressure status of the factory. This is also responsible for determining if a job can be loadshed.

Required Methods

fn is_discardable(&self, job: &TKey) -> bool

Determine if this job can be discarded under load.

Returns true if the job can be discarded, false otherwise.

fn get_priority(&self, job: &TKey) -> Option<TPriority>

Retrieve the job’s priority.

Returns None if the job does not have a priority, [Some(TPriority)] otherwise.