ractor::factory

Module ratelim

Rate limiting protocols for factory routers

Structs

LeakyBucketRateLimiter

A basic leaky-bucket rate limiter. This is a synchronous implementation with no interior locking since it’s only used by the RateLimitedRouter uniquely and doesn’t share its state

LeakyBucketRateLimiterBuilder

Use builder syntax to set the required parameters and finish by calling the method Self::build().

RateLimitedRouter

A generic struct which wraps the message router and adds support for a rate-limiting implementation to rate limit jobs processed by the factory. This handles the plubming around wrapping a rate limited message router

RateLimitedRouterBuilder

Use builder syntax to set the required parameters and finish by calling the method Self::build().

Constants

MAX_LB_BALANCE

The maximum supported balance for leaky bucket rate limiting.

Traits

RateLimiter

A basic trait which allows controlling rate limiting of message routing