ractor::factory::ratelim

Trait RateLimiter

pub trait RateLimiter: State {
    // Required methods
    fn check(&mut self) -> bool;
    fn bump(&mut self);
}

A basic trait which allows controlling rate limiting of message routing

Required Methods

fn check(&mut self) -> bool

Check if we have not violated the rate limiter

Returns false if we’re in violation and should start rate-limiting traffic true otherwise

fn bump(&mut self)

Bump the rate limit internal counter, as we’ve routed a message to a worker