ractor::factory

Module routing

Routing protocols for Factories

Structs

CustomRouting

Factory will dispatch to workers based on a custom hash function.

KeyPersistentRouting

Factory will select worker by hashing the job’s key. Workers will have jobs placed into their incoming message queue’s

QueuerRouting

Factory will dispatch job to first available worker. Factory will maintain shared internal queue of messages

RoundRobinRouting

Factory will dispatch to the next worker in order.

StickyQueuerRouting

Factory will dispatch jobs to a worker that is processing the same key (if any). Factory will maintain shared internal queue of messages.

Enums

RouteResult

The possible results from a routing operation.

Traits

CustomHashFunction

Custom hashing behavior for factory routing to workers

Router

A routing mode controls how a request is routed from the factory to a designated worker