ractor::factory::stats

Trait FactoryStatsLayer

pub trait FactoryStatsLayer:
    Send
    + Sync
    + 'static {
    // Required methods
    fn factory_ping_received(&self, factory: &str, sent: Instant);
    fn worker_ping_received(&self, factory: &str, elapsed: Duration);
    fn new_job(&self, factory: &str);
    fn job_completed(&self, factory: &str, options: &JobOptions);
    fn job_discarded(&self, factory: &str);
    fn job_rate_limited(&self, factory: &str);
    fn job_ttl_expired(&self, factory: &str, num_removed: usize);
    fn record_queue_depth(&self, factory: &str, depth: usize);
    fn record_processing_messages_count(&self, factory: &str, count: usize);
    fn record_worker_count(&self, factory: &str, count: usize);
    fn record_queue_limit(&self, factory: &str, count: usize);

    // Provided method
    fn record_in_flight_messages_count(&self, factory: &str, count: usize) { ... }
}

A wrapper over whatever stats collection a user wishes to utilize in the super::Factory.

Required Methods

fn factory_ping_received(&self, factory: &str, sent: Instant)

Called when a factory ping has been received, marking the duration between when it was sent and now.

Measures ping latency on the factory

fn worker_ping_received(&self, factory: &str, elapsed: Duration)

Called when a worker replies to a ping request from the factory, measuring the duration between the time the ping was sent and the factory processed the ping response.

Measures worker “free” latency and this metric relates to identification of “stuck” or slow workers.

fn new_job(&self, factory: &str)

Called for each new incoming job

fn job_completed(&self, factory: &str, options: &JobOptions)

Called when a job is completed to report factory processing time, worker processing time, total processing time, and job count

From these metrics you can derive

Time in factory’s queue = factory_job_processing_latency_usec - worker_job_processing_latency_usec

Time in worker’s queue + being processed by worker = worker_job_processing_latency_usec

Total time since submission = job_processing_latency_usec

fn job_discarded(&self, factory: &str)

Called when a job is discarded

fn job_rate_limited(&self, factory: &str)

Called when a job is rate limited

fn job_ttl_expired(&self, factory: &str, num_removed: usize)

Called when jobs TTL timeout in the factory’s queue

fn record_queue_depth(&self, factory: &str, depth: usize)

Fixed-period recording of the factory’s queue depth

fn record_processing_messages_count(&self, factory: &str, count: usize)

Fixed-period recording of the factory’s number of processed messages

fn record_worker_count(&self, factory: &str, count: usize)

Fixed-period recording of the factory’s number of workers

fn record_queue_limit(&self, factory: &str, count: usize)

Fixed-period recording of the factory’s maximum allowed queue size

Provided Methods

fn record_in_flight_messages_count(&self, factory: &str, count: usize)

Fixed-period recording of the factory’s in-flight message count (processing + queued)

Default empty implemention for backwards compatibility