ractor::factory

Struct UpdateSettingsRequest

pub struct UpdateSettingsRequest<TKey, TMsg>
where
    TKey: JobKey,
    TMsg: Message,
{
    pub discard_handler: Option<Option<Arc<dyn DiscardHandler<TKey, TMsg>>>>,
    pub discard_settings: Option<DiscardSettings>,
    pub dead_mans_switch: Option<Option<DeadMansSwitchConfiguration>>,
    pub capacity_controller: Option<Option<Box<dyn WorkerCapacityController>>>,
    pub lifecycle_hooks: Option<Option<Box<dyn FactoryLifecycleHooks<TKey, TMsg>>>>,
    pub stats: Option<Option<Arc<dyn FactoryStatsLayer>>>,
    pub worker_count: Option<usize>,
}

The settings to change for an update request to the factory at runtime.

Note: A value of Some(..) means that the internal value should be updated inside the factory’s state. For values which are originally optional to the factory, we use Option<Option<T>>, so if you want to UNSET the value, it would be Some(None).

Fields
discard_handler: Option<Option<Arc<dyn DiscardHandler<TKey, TMsg>>>>

The discard handler callback

discard_settings: Option<DiscardSettings>

The discard settings

dead_mans_switch: Option<Option<DeadMansSwitchConfiguration>>

Dead-man’s switch settings

capacity_controller: Option<Option<Box<dyn WorkerCapacityController>>>

Capacity controller

lifecycle_hooks: Option<Option<Box<dyn FactoryLifecycleHooks<TKey, TMsg>>>>

Lifecycle hooks

stats: Option<Option<Arc<dyn FactoryStatsLayer>>>

Statistics layer

worker_count: Option<usize>

The worker count

Implementations

impl<TKey, TMsg> UpdateSettingsRequest<TKey, TMsg>
where TKey: JobKey, TMsg: Message,

pub fn builder() -> UpdateSettingsRequestBuilder<TKey, TMsg>

Create an instance of UpdateSettingsRequest using the builder syntax