ractor::factory

Module worker

Factory worker properties

Structs

DeadMansSwitchConfiguration

The configuration for the dead-man’s switch functionality

DeadMansSwitchConfigurationBuilder

Use builder syntax to set the required parameters and finish by calling the method Self::build().

WorkerProperties

Properties of a worker

WorkerStartContext

Startup context data (Arguments) which are passed to a worker on start

Enums

WorkerMessage

Message to a worker

Traits

Worker

A factory worker trait, which is a basic wrapper around actor logic, with predefined type information specific to workers

WorkerBuilder

The super::Factory is responsible for spawning workers and re-spawning workers under failure scenarios. This means that it needs to understand how to build workers. The WorkerBuilder trait is used by the factory to construct new workers when needed.

WorkerCapacityController

Controls the size of the worker pool by dynamically growing/shrinking the pool to requested size