ractor

Macro call

macro_rules! call {
    ($actor:expr, $msg:expr) => { ... };
    ($actor:expr, $msg:expr, $($args:expr),*) => { ... };
}

call!: Perform an infinite-time remote procedure call to an crate::Actor

$actor - The actor to call

$msg - The message builder which takes in a crate::port::RpcReplyPort and emits a message which the actor supports

$args - (optional) Variable length arguments which will PRECEDE the reply channel when constructing the message payload

Returns [Ok(_)] with the result on successful RPC or [Err(crate::RactorErr)] on failure Example usage (without the cluster feature)

use ractor::{call, Actor, RpcReplyPort, ActorRef, ActorProcessingErr};
struct TestActor;
enum MessageFormat {
    TestRpc(String, RpcReplyPort<String>),
}
#[cfg(feature = "cluster")]
impl ractor::Message for MessageFormat {}

#[cfg_attr(feature = "async-trait", ractor::async_trait)]
impl Actor for TestActor {
    type Msg = MessageFormat;
    type Arguments = ();
    type State = ();

    async fn pre_start(&self, _this_actor: ActorRef<Self::Msg>, _: ()) -> Result<Self::State, ActorProcessingErr> {
        Ok(())
    }

    async fn handle(
        &self,
        _this_actor: ActorRef<Self::Msg>,
        message: Self::Msg,
        _state: &mut Self::State,
    ) -> Result<(), ActorProcessingErr> {
        match message {
            Self::Msg::TestRpc(arg, reply) => {
                // An error sending means no one is listening anymore (the receiver was dropped),
                // so we should shortcut the processing of this message probably!
                if !reply.is_closed() {
                    let _ = reply.send(arg);
                }
            }
        }
        Ok(())
    }
}

#[tokio::main]
async fn main() {
    let (actor, handle) = Actor::spawn(None, TestActor, ()).await.unwrap();
    let result = call!(actor, MessageFormat::TestRpc, "Something".to_string()).unwrap();
    assert_eq!(result, "Something".to_string());
    actor.stop(None);
    handle.await.unwrap();
}