ractor

Module message

Message trait definition for inter-actor messaging. Additionally with the cluster feature, it controls serialization logic for over-the-wire inter-actor communications

Structs

BoxedDowncastErr

An error downcasting a boxed item to a strong type

BoxedMessage

A “boxed” message denoting a strong-type message but generic so it can be passed around without type constraints

Traits

Message

Message type for an actor. Generally an enum which muxes the various types of inner-messages the actor supports