ractor

Module pg

Process groups (PG) are named groups of actors with a friendly name which can be used for retrieval of the process groups. Then within the group, either a random actor (for dispatch) can be selected or the whole group (broadcast), or a subset (partial-broadcast) can have a message sent to them. Common operations are to (a) upcast the group members to a strong-type’d actor then dispatch a message with crate::call or crate::cast.

Process groups can also be monitored for changes with calling monitor to subscribe to changes and demonitor to unsubscribe. Subscribers will receive process group change notifications via a SupervisionEvent called on the supervision port of the crate::Actor

Inspired from Erlang’s pg module

Examples

use ractor::pg;
use ractor::Actor;
use ractor::ActorProcessingErr;
use ractor::ActorRef;

struct ExampleActor;

#[cfg_attr(feature = "async-trait", ractor::async_trait)]
impl Actor for ExampleActor {
    type Msg = ();
    type State = ();
    type Arguments = ();

    async fn pre_start(
        &self,
        _myself: ActorRef<Self::Msg>,
        _args: Self::Arguments,
    ) -> Result<Self::State, ActorProcessingErr> {
        println!("Starting");
        Ok(())
    }
}

#[tokio::main]
async fn main() {
    let (actor, handle) = Actor::spawn(None, ExampleActor, ())
        .await
        .expect("Failed to startup dummy actor");
    let group = "the_group".to_string();

    // Join the actor to a group. This is also commonly done in `pre_start` or `post_start`
    // of the actor itself without having to do it externally by some coordinator
    pg::join(group.clone(), vec![actor.get_cell()]);
    // Retrieve the pg group membership
    let members = pg::get_members(&group);
    // Send a message to the up-casted actor
    let the_actor: ActorRef<()> = members.get(0).unwrap().clone().into();
    ractor::cast!(the_actor, ()).expect("Failed to send message");

    // wait for actor exit
    actor.stop(None);
    handle.await.unwrap();

    // The actor will automatically be removed from the group upon shutdown.
    let members = pg::get_members(&group);
    assert_eq!(members.len(), 0);
}

Structs

ScopeGroupKey

Represents the combination of a ScopeName and a GroupName that uniquely identifies a specific group in a specific scope

Enums

GroupChangeMessage

Represents a change in a process group’s membership

Constants

ALL_GROUPS_NOTIFICATION

Key to monitor all of the groups in a scope

ALL_SCOPES_NOTIFICATION

Key to monitor all of the scopes

DEFAULT_SCOPE

Key to set the default scope

Functions

demonitor

Unsubscribes the provided crate::Actor for updates from the group in default scope

demonitor_scope

Unsubscribes the provided crate::Actor from the scope for updates

get_local_members

Returns all actors running on the local node in the group group in the default scope.

get_members

Returns all the actors running on any node in the group group in the default scope.

get_scoped_local_members

Returns all actors running on the local node in the group group in scope scope

get_scoped_members

Returns all the actors running on any node in the group group in the scope scope.

join

Join actors to the group group in the default scope

join_scoped

Join actors to the group group within the scope scope

leave

Leaves the specified crate::Actors from the PG group in the default scope

leave_scoped

Leaves the specified crate::Actors from the PG group within the scope scope

monitor

Subscribes the provided crate::Actor to the group in the default scope for updates

monitor_scope

Subscribes the provided crate::Actor to the scope for updates

which_groups

Return a list of all known groups

which_scoped_groups

Returns a list of all known groups in scope scope

which_scopes

Returns a list of all known scopes

which_scopes_and_groups

Returns a list of all known scope-group combinations.