ractor::port

Module output

Output ports for publish-subscribe notifications between actors

This notion extends beyond traditional actors in this that is a publish-subscribe mechanism we’ve added in ractor. Output ports are ports which can have messages published to them which are automatically forwarded to downstream actors waiting for inputs. They optionally have a message transformer attached to them to convert them to the appropriate message type

There are two different implementation. If the feature output-port-v2 is not specified the implementation use a broadcast channel that is limited to 10 messages successively sent for each susbscribed actor. That means that if 10 messages are sent to the output port successively there is a high probably that any subscriber receive all messages.

The new implementation that is accessible using output-port-v2 use a fan-out task that distributes message to all subscriber ensuring that all messages are received by all subscribers

Structs

OutputPort

An OutputPort is a publish-subscribe mechanism for connecting actors together. It allows actors to emit messages without knowing which downstream actors are subscribed.

Traits

OutputMessage

Output messages, since they need to be replicated, require Clone in addition to the base Message constraints

OutputPortSubscriberTrait

A trait for subscribing to an OutputPort

Type Aliases

OutputPortSubscriber

Represents a boxed ActorRef subscriber capable of handling messages from a publisher via an OutputPort, employing a publish-subscribe pattern to decouple message broadcasting from handling. For a subscriber ActorRef to function as an OutputPortSubscriber<T>, its message type must implement From<T> to convert the published message type to its own message format.