ractor::port

Struct RpcReplyPort

pub struct RpcReplyPort<TMsg> { /* private fields */ }

A remote procedure call’s reply port. Wrapper of concurrency::OneshotSender with a consistent error type

Implementations

impl<TMsg> RpcReplyPort<TMsg>

pub fn get_timeout(&self) -> Option<Duration>

Read the timeout of this RPC reply port

Returns [Some(concurrency::Duration)] if a timeout is set, None otherwise

pub fn send(self, msg: TMsg) -> Result<(), MessagingErr<TMsg>>

Send a message to the Rpc reply port. This consumes the port

msg - The message to send

Returns [Ok(())] if the message send was successful, [Err(MessagingErr)] otherwise

pub fn is_closed(&self) -> bool

Determine if the port is closed (i.e. the receiver has been dropped)

Returns true if the receiver has been dropped and the channel is closed, this means sends will fail, false if channel is open and receiving messages