ractor::rpc::call_result

Enum CallResult

pub enum CallResult<TResult> {
    Success(TResult),
    Timeout,
    SenderError,
}

The result from a crate::rpc::call operation

Variants

Success(TResult)

Success, with the result

Timeout

Timeout

SenderError

The transmission channel was dropped without any message(s) being sent

Implementations

impl<T> CallResult<T>

pub fn is_success(&self) -> bool

Determine if the CallResult is a CallResult::Success

pub fn is_timeout(&self) -> bool

Determine if the CallResult is a CallResult::Timeout

pub fn is_send_error(&self) -> bool

Determine if the CallResult is a CallResult::SenderError

pub fn unwrap(self) -> T

Unwrap a CallResult, panicking on any non-success

pub fn expect(self, msg: &'static str) -> T

Unwrap a CallResult, panicking on non-succcess with the specified message

pub fn unwrap_or(self, default: T) -> T

Unwrap the CallResult or give a default value

pub fn unwrap_or_else<F>(self, f: F) -> T
where F: FnOnce() -> T,

Returns the CallResult’s success result or computes the closure

pub fn success_or<E>(self, err: E) -> Result<T, E>

Transforms the CallResult to a Result mapping Success(t) to Ok(t) and all else to Err(err)

pub fn success_or_else<E, F>(self, err: F) -> Result<T, E>
where F: FnOnce() -> E,

Transforms the CallResult to a Result mapping Success(t) to Ok(t) and all else to Err(err())

pub fn map<O, F>(self, mapping: F) -> CallResult<O>
where F: FnOnce(T) -> O,

Maps the success value of the CallResult to another type

pub fn map_or<O, F>(self, default: O, mapping: F) -> O
where F: FnOnce(T) -> O,

Maps the success value of the CallResult to another type or returns the default value

pub fn map_or_else<D, O, F>(self, default: D, mapping: F) -> O
where F: FnOnce(T) -> O, D: FnOnce() -> O,

Maps the success value of the CallResult to another type or returns the default function result